move test_update_bots_endpoint tests -> test_bb_scopes_updates
